Adds debug/fetch_cas
