+++ conductor_api/src/conductor/debug.rs
@@ -1,11 +1,14 @@
 use conductor::Conductor;
 use holochain_core_types::error::HolochainError;
 use holochain_core::state_dump::StateDump;
+use holochain_persistence_api::cas::content::Address;
 
 pub trait ConductorDebug {
     fn running_instances(&self) -> Result<Vec<String>, HolochainError>;
     fn state_dump_for_instance(&self, instance_id: &String)
         -> Result<StateDump, HolochainError>;
+    fn get_type_and_content_from_cas(&self, address: &Address, instance_id: &String)
+        -> Result<(String, String), HolochainError>;
 }
 
 impl ConductorDebug for Conductor {
@@ -17,5 +20,11 @@ impl ConductorDebug for Conductor {
         -> Result<StateDump, HolochainError> {
         let hc = self.instances.get(instance_id)?;
         Ok(hc.read().unwrap().get_state_dump()?)
+    }
+
+    fn get_type_and_content_from_cas(&self, address: &Address, instance_id: &String)
+        -> Result<(String, String), HolochainError> {
+        let hc = self.instances.get(instance_id)?;
+        Ok(hc.read().unwrap().get_type_and_content_from_cas(address)?)
     }
 }
